Rename IdGenerators namespace to StronglyTypedIds and move StronglyTypedDocumentProcessor into it
