+++ application/account-management/Core/Features/Signups/Domain/Signup.cs
@@ -1,6 +1,6 @@
 using JetBrains.Annotations;
 using PlatformPlatform.SharedKernel.Domain;
-using PlatformPlatform.SharedKernel.IdGenerators;
+using PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 namespace PlatformPlatform.AccountManagement.Features.Signups.Domain;
