+++ application/account-management/Tests/FakerExtensions.cs
@@ -1,5 +1,5 @@
 using Bogus;
-using PlatformPlatform.SharedKernel.IdGenerators;
+using PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 namespace PlatformPlatform.AccountManagement.Tests;
