+++ application/shared-kernel/SharedKernel/Authentication/RefreshTokenId.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.SharedKernel.IdGenerators;
+using PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 namespace PlatformPlatform.SharedKernel.Authentication;
