+++ application/shared-kernel/SharedKernel/Domain/TenantId.cs
@@ -1,5 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
-using PlatformPlatform.SharedKernel.IdGenerators;
+using PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 namespace PlatformPlatform.SharedKernel.Domain;
