+++ application/shared-kernel/SharedKernel/Domain/UserId.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.SharedKernel.IdGenerators;
+using PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 namespace PlatformPlatform.SharedKernel.Domain;
