+++ application/shared-kernel/SharedKernel/StronglyTypedIds/IdPrefixAttribute.cs
@@ -1,4 +1,4 @@
-namespace PlatformPlatform.SharedKernel.IdGenerators;
+namespace PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 [AttributeUsage(AttributeTargets.Class)]
 public sealed class IdPrefixAttribute(string prefix) : Attribute
