+++ application/shared-kernel/SharedKernel/StronglyTypedIds/StronglyTypedId.cs
@@ -1,4 +1,4 @@
-namespace PlatformPlatform.SharedKernel.IdGenerators;
+namespace PlatformPlatform.SharedKernel.StronglyTypedIds;
 
 /// <summary>
 ///     StronglyTypedId is an abstract record type for creating strongly typed IDs with a specified value type. It makes
