+++ application/shared-kernel/Tests/Entities/EntityTests.cs
@@ -1,6 +1,6 @@
 using FluentAssertions;
 using PlatformPlatform.SharedKernel.Domain;
-using PlatformPlatform.SharedKernel.IdGenerators;
+using PlatformPlatform.SharedKernel.StronglyTypedIds;
 using Xunit;
 
 namespace PlatformPlatform.SharedKernel.Tests.Entities;
